Fixed Multiple issues related to disconnected users, implemented automatic loss when someone disconnects from a match for longer than 10s, implemented special abilities (railshot - bugged, speedup works fine), Implemented UI for matchseek menu (Versus Mode)

+++ Backend/UserManagerBackend/UserManagement/UserDataManagement/models.py
@@ -29,7 +29,7 @@ class User(models.Model):
     matcheslost = models.IntegerField(db_column='MatchesLost', default=0)  
     xp = models.IntegerField(db_column='XP', default=0) 
     rank = models.IntegerField(db_column='Rank', default=1)
-    level = models.IntegerField(db_column='level', default=0)
+    level = models.DecimalField(db_column='level', default=0, decimal_places=2, max_digits=2)
     utitle = models.CharField(db_column='uTitle', max_length=16, default="")
     udiscordid = models.CharField(db_column='uDiscordID', max_length=16, default="")
     utournamentsplayed = models.IntegerField(db_column='uTournamentsPlayed', default=0)
